Move some checks from paramfits to the main checks
And use them for the tables, where they are needed in order to prevent
pandas from merging the index.
