+++ validphys2/src/validphys/results.py
@@ -19,7 +19,7 @@ from reportengine.checks import require_one, remove_outer, check_not_empty, make
 from reportengine.table import table
 from reportengine import collect
 
-from validphys.checks import assert_use_cuts_true, check_pdf_is_montecarlo
+from validphys.checks import assert_use_cuts_true, check_pdf_is_montecarlo, check_speclabels_different
 from validphys.core import DataSetSpec, PDF, ExperimentSpec
 from validphys.calcutils import all_chi2, central_chi2, calc_chi2, calc_phi, bootstrap_values
 
@@ -663,6 +663,7 @@ def fits_experiments_chi2_table(fits, fits_experiments, fits_experiment_chi2_dat
     return res
 
 @table
+@check_speclabels_different
 def dataspecs_experiments_chi2_table(dataspecs_speclabel, dataspecs_experiments,
                                      dataspecs_experiment_chi2_data,
                                      per_point_data:bool=True):
@@ -712,6 +713,7 @@ def fits_datasets_chi2_table(fits, fits_experiments, fits_chi2_data,
     return pd.concat(dfs, axis=1)
 
 @table
+@check_speclabels_different
 def dataspecs_datasets_chi2_table(dataspecs_speclabel, dataspecs_experiments,
                                   dataspecs_chi2_data, per_point_data:bool=True):
     """Same as fits_datasets_chi2_table but for arbitrary dataspecs."""
